Add custom slider
This is a slider inspired by those used in Blender.

The control is a draggable slider displaying the current value as its
background. Clicking the control brings up a LineEdit for manual value
input.

When dragging the slider, hold CTRL to round to 0.1 increments or hold
SHIFT to round to 0.01 increments. Holding both CTRL and SHIFT rounds to
0.001 increments.